Refactor for release health query re-use.
