+++ pkg/db/query/payload_queries.go
@@ -1,0 +1,77 @@
+package query
+
+import (
+	"gorm.io/gorm"
+
+	"github.com/openshift/sippy/pkg/db/models"
+)
+
+// GetLastAcceptedByArchitectureAndStream returns the last accepted payload for each architecture/stream combo.
+func GetLastAcceptedByArchitectureAndStream(db *gorm.DB, release string) ([]models.ReleaseTag, error) {
+	results := make([]models.ReleaseTag, 0)
+
+	result := db.Raw(`SELECT
+						DISTINCT ON
+							(architecture, stream)
+							*
+						FROM
+							release_tags
+						WHERE
+							release = ?
+						AND
+							phase = 'Accepted'
+						ORDER BY
+							architecture, stream, release_time desc`, release).Scan(&results)
+
+	if result.Error != nil {
+		return nil, result.Error
+	}
+
+	return results, nil
+}
+
+// GetLastPayloadStatus returns the most recent payload status for an architecture/stream combination,
+// as well as the count of how many of the last payloads had that status (e.g., when this returns
+// Rejected, 5 -- it means the last 5 payloads were rejected.
+func GetLastPayloadStatus(db *gorm.DB, architecture, stream, release string) (string, int, error) {
+	count := struct {
+		Phase string `gorm:"column:phase"`
+		Count int    `gorm:"column:count"`
+	}{}
+
+	result := db.Raw(`
+		WITH releases AS
+			(
+				SELECT
+					ROW_NUMBER() OVER(ORDER BY release_time desc) AS id,
+					phase
+				FROM
+					release_tags
+				WHERE
+					architecture = ? AND stream = ? AND release = ?
+			),
+		changes AS
+			(
+				SELECT
+					*,
+					CASE WHEN LAG(phase) OVER(ORDER BY id) = phase THEN 0 ELSE 1 END AS change
+				FROM
+					releases
+			),
+		groups AS
+			(
+				SELECT
+					*,
+					SUM(change) OVER(ORDER BY id) AS group FROM changes
+			)
+		SELECT
+			phase, COUNT(phase)
+		FROM
+			groups
+		WHERE
+			groups.group = 1
+		GROUP BY
+			phase`, architecture, stream, release).Scan(&count)
+
+	return count.Phase, count.Count, result.Error
+}
